pyproject.toml
[tool.poetry]
name = "study"
version = "0.0.1"
description = ""
authors = ["shun198 <shunhiroseluvmri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
drf-nested-routers = "^0.93.4"
django-extensions = "^3.2.1"
redis = "^4.3.4"
django-filter = "^22.1"
django-debug-toolbar = "^3.7.0"
Authlib = "^1.1.0"
django-cors-headers = "^3.13.0"
pytest = "^7.2.0"
pytest-custom-exit-code = "^0.3.0"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.6"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.group.dev.dependencies]
Django = "^4.1.2"
mysqlclient = "^2.1.1"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
